fix(realworld): guard dummy step + rename franky joint env file
* DualFrankaFrankyEnv.step: hold the ctrls list reference inside the
  `if not is_dummy:` block. In dummy mode `_setup_hardware` is skipped
  and `_left_ctrl` / `_right_ctrl` don't exist, so the prior unconditional
  unpack would AttributeError. Per reviewer.
* Rename `dual_franka_joint_env.py` → `dual_franka_franky_joint_env.py`
  to make the franky / libfranka heritage obvious. Class name
  `DualFrankaJointEnv` and registered env id `DualFrankaJointEnv-v1`
  are unchanged so no external import or YAML reference moves. Per
  reviewer.
